Remove alert about missing support for highp precision in shaders
On Firefox, this alert would show because minRange for HIGH_INT returned by getShaderPrecisionFormat() is reported as 24 instead of 31 (on Linux). Which does not make sense, because the minimum requirement in WebGL 2.0 is 31. So if WebGL 2.0 context is supported in the browser, we now just assume that float32 and (u)int32 is available in both vertex and fragment shaders.