Implement SDF-based rendering of marker shapes (WIP)
Rendering markers with signed distance fields (SDFs) means that we can now also change the outline/stroke width, and show non-filled markers with outlines only. These features are currently not exposed in the UI, but can be tested by changing the new options in dataUtils or glUtils from the console.

Known limitations/regressions:
- Thicker outlines are clipped for some shapes. This is caused by the scaling of UV coordinates in the fragment shader, so adding a fix for this would likely require users to also updating the marker scale factor in their TissUUmaps projects.
- The Gaussian shape is currently broken.