Add cluster information to improve performance of region rendering
This adds basic clustering of bounding boxes for paths in the edge data. Should reduce the number of bounding box checks compared to using pivot split, as long as there is some spatial coherence between the region paths in the dataset. For now, the clustering is done in a separate pass, but could also be combined with the creation of the edge lists to make update after editing faster.