Add code for splitting long clusters in the region edge lists
When adding clusters to edge lists, a new cluster will now be started if the current cluster's bounding box exceeds a threshold determined by the scanline width (the image width) of a layer. This avoids creating long clusters, which allows us to also increase the cluster size. The performance gain should be very noticable for large region datasets.