Remove index buffer to make marker sorting work also with instanced drawing
glUtils.loadMarkers now generates a vertex buffer where data is already
sorted, instead of as before an unsorted vertex buffer + a sorted index
buffer. This will be worse for caching when the sorted and unsorted
order differs a lot (compared to just sorting the index list), but is
needed to make the marker sorting work with the instanced drawing.