Add code for rendering markers with absolute size (size in image pixels)
This adds a new per UID option (currently only in glUtils) for toggling if markers should be rendered with absolute size, i.e. in units matching pixels in the corresponding image layer. Note: the global marker size slider will not affect the scaling when this is enabled (but that is intentional).